+++ .gitignore
@@ -1,1 +1,2 @@
 flask_session
+_pycache_
